+++ src/ts/undo/IRUndo.ts
@@ -14,23 +14,28 @@ class IRUndo {
     private lastText: string;
     private hasUndo: boolean;
 
-    constructor(vditor: IVditor) {
+    constructor() {
+        this.redoStack = [];
+        this.undoStack = [];
         // @ts-ignore
         this.dmp = new DiffMatchPatch();
-        this.clearStack(vditor);
+        this.lastText = "";
+        this.hasUndo = false;
     }
 
     public clearStack(vditor: IVditor) {
         this.redoStack = [];
         this.undoStack = [];
         this.lastText = "";
-        this.hasUndo = false; 
-        if (vditor.toolbar) {
-            this.resetIcon(vditor);
-        }
+        this.hasUndo = false;
+        this.resetIcon(vditor);
     }
 
     public resetIcon(vditor: IVditor) {
+        if (!vditor.toolbar) {
+            return;
+        }
+
         if (this.undoStack.length > 1) {
             enableToolbar(vditor.toolbar.elements, ["undo"]);
         } else {